Update loginUtils file
